refactor: solution과 problem으로 DTO 구분
Related to: #160
